+++ application/client/root/src/routing/appUrls.ts
@@ -1,0 +1,5 @@
+import { createObjectNiceWebRoutes } from 'nice-web-routes';
+
+export const appUrls = createObjectNiceWebRoutes()({
+  some: {},
+});
