+++ application/client/root/src/routing/index.ts
@@ -1,0 +1,1 @@
+export * from './appUrls';
